Reworked toggling classes for selects and radio buttons, refactored radio setup and checkbox dispatch
If a `data-add-class` on a select or radio button does not include a
selector and there is a `data-show` property it will add the class name
to the currently visible element and remove it from the hidden ones.

For example a select with `data-add-class='active'` will add `.active`
to the `div.menu-1` when an option with `data-show='.menu-1'` is chosen.

The dispatch function now properly handles forced values, this allowed
me to rewrite the toggleCheckbox function to use dispatch.

Radio buttons now behave like select options, calculating their data-hide
at runtime based on their form siblings with a matching name attribute.